requirements.txt
openai==1.60.2
pymongo==4.11
python-dotenv==1.0.1
edge_tts==7.0.0
debugpy==1.8.0
pika==1.3.2
beanie==1.29.0
discord.py==2.5.2
pydantic==2.11
PyPDF2==3.0.1
pydub==0.25.1
pytest==8.4.1
